HBX-1639: Extract used methods of 'org.hibernate.tool.hbm2x.StringUtils' into 'org.hibernate.tool.internal.util'
